Centralize experimental API seriliazers
Problem was that, there were two PackageCategorySerializer classes
Solution, remove one and move all serializers to one place
